Move the schema to a new module
